add: タイムライン画面のUIのリファクタリングをしました
