+++ lib/ui/screen/time_line/time_line_screen.dart
@@ -5,18 +5,17 @@ import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:food_gram_app/core/admob/services/admob_open.dart';
 import 'package:food_gram_app/core/admob/tracking/ad_tracking_permission.dart';
 import 'package:food_gram_app/core/local/force_update_checker.dart';
-import 'package:food_gram_app/core/model/tag.dart';
 import 'package:food_gram_app/core/purchase/services/revenue_cat_service.dart';
 import 'package:food_gram_app/core/supabase/post/providers/block_list_provider.dart';
 import 'package:food_gram_app/core/supabase/post/providers/post_stream_provider.dart';
 import 'package:food_gram_app/core/utils/helpers/dialog_helper.dart';
 import 'package:food_gram_app/gen/assets.gen.dart';
-import 'package:food_gram_app/gen/l10n/l10n.dart';
 import 'package:food_gram_app/router/router.dart';
+import 'package:food_gram_app/ui/component/app_category_item.dart';
 import 'package:food_gram_app/ui/component/app_error_widget.dart';
 import 'package:food_gram_app/ui/component/app_floating_button.dart';
 import 'package:food_gram_app/ui/component/app_list_view.dart';
-import 'package:gap/gap.dart';
+import 'package:food_gram_app/ui/component/app_story_widget.dart';
 import 'package:go_router/go_router.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
@@ -119,90 +118,30 @@ class RestaurantCategoryScreen extends HookConsumerWidget {
     final selectedCategoryName = useState('');
     final postState =
         ref.watch(postStreamByCategoryProvider(selectedCategoryName.value));
-    final l10n = L10n.of(context);
-    return Column(
-      children: [
-        SizedBox(
-          height: 60,
-          child: ListView.builder(
-            scrollDirection: Axis.horizontal,
-            itemCount: foodCategory.length + 1,
-            itemBuilder: (context, index) {
-              if (index == 0) {
-                return GestureDetector(
-                  onTap: () {
-                    selectedCategoryName.value = '';
-                  },
-                  child: Container(
-                    margin: EdgeInsets.symmetric(horizontal: 6, vertical: 8),
-                    padding: EdgeInsets.symmetric(horizontal: 12, vertical: 4),
-                    decoration: BoxDecoration(
-                      color: selectedCategoryName.value.isEmpty
-                          ? Colors.black
-                          : Colors.white,
-                      borderRadius: BorderRadius.circular(20),
-                      border: Border.all(),
-                    ),
-                    child: Row(
-                      children: [
-                        Text('🍽️', style: TextStyle(fontSize: 18)),
-                        SizedBox(width: 4),
-                        Text(
-                          l10n.foodCategoryAll,
-                          style: TextStyle(
-                            color: selectedCategoryName.value.isEmpty
-                                ? Colors.white
-                                : Colors.black,
-                            fontWeight: FontWeight.bold,
-                          ),
-                        ),
-                      ],
-                    ),
-                  ),
-                );
-              }
-              final entry = foodCategory.entries.elementAt(index - 1);
-              final categoryName = entry.key;
-              final foodEmojis = entry.value;
-              final displayIcon = foodEmojis.isNotEmpty ? foodEmojis[0] : '🍽️';
-              final isSelected = selectedCategoryName.value == categoryName;
-              return GestureDetector(
-                onTap: () {
-                  selectedCategoryName.value = categoryName;
-                },
-                child: Container(
-                  margin: EdgeInsets.symmetric(horizontal: 6, vertical: 8),
-                  padding: EdgeInsets.symmetric(horizontal: 8, vertical: 4),
-                  decoration: BoxDecoration(
-                    color: isSelected ? Colors.black : Colors.white,
-                    borderRadius: BorderRadius.circular(20),
-                    border: Border.all(),
-                  ),
-                  child: Row(
-                    children: [
-                      Text(displayIcon, style: TextStyle(fontSize: 24)),
-                      Gap(4),
-                      Text(
-                        l10nCategory(categoryName, l10n),
-                        style: TextStyle(
-                          color: isSelected ? Colors.white : Colors.black,
-                          fontWeight: FontWeight.bold,
-                        ),
-                      ),
-                    ],
-                  ),
+    return NestedScrollView(
+      headerSliverBuilder: (context, innerBoxIsScrolled) {
+        return [
+          SliverToBoxAdapter(
+            child: Column(
+              children: [
+                SizedBox(
+                  width: double.infinity,
+                  height: 100,
+                  child: AppStoryWidget(data: postState.value ?? []),
                 ),
-              );
-            },
+                AppCategoryItem(selectedCategoryName: selectedCategoryName),
+              ],
+            ),
           ),
-        ),
-        Expanded(
-          child: FoodListView(
-            state: postState,
-            isRestaurant: true,
-          ),
-        ),
-      ],
+        ];
+      },
+      body: NotificationListener<ScrollNotification>(
+        onNotification: (ScrollNotification notification) {
+          // スクロール通知を処理
+          return false;
+        },
+        child: FoodListView(state: postState, isRestaurant: true),
+      ),
     );
   }
 }
@@ -249,47 +188,4 @@ class FoodListView extends ConsumerWidget {
       ),
     );
   }
-}
-
-String l10nCategory(String categoryName, L10n l10n) {
-  String displayName;
-  switch (categoryName) {
-    case 'Noodles':
-      displayName = l10n.foodCategoryNoodles;
-      break;
-    case 'Meat':
-      displayName = l10n.foodCategoryMeat;
-      break;
-    case 'Fast Food':
-      displayName = l10n.foodCategoryFastFood;
-      break;
-    case 'Rice Dishes':
-      displayName = l10n.foodCategoryRiceDishes;
-      break;
-    case 'Seafood':
-      displayName = l10n.foodCategorySeafood;
-      break;
-    case 'Bread':
-      displayName = l10n.foodCategoryBread;
-      break;
-    case 'Sweets & Snacks':
-      displayName = l10n.foodCategorySweetsAndSnacks;
-      break;
-    case 'Fruits':
-      displayName = l10n.foodCategoryFruits;
-      break;
-    case 'Vegetables':
-      displayName = l10n.foodCategoryVegetables;
-      break;
-    case 'Beverages':
-      displayName = l10n.foodCategoryBeverages;
-      break;
-    case 'Others':
-      displayName = l10n.foodCategoryOthers;
-      break;
-    default:
-      displayName = categoryName;
-  }
-
-  return displayName;
 }
